+++ tests/HandlerTest.php
@@ -6,14 +6,21 @@ namespace Tests\Innmind\Signals;
 use Innmind\Signals\{
     Handler,
     Signal,
+    Async\Interceptor,
 };
-use PHPUnit\Framework\TestCase;
+use Innmind\BlackBox\{
+    PHPUnit\BlackBox,
+    PHPUnit\Framework\TestCase,
+    Set,
+};
 
 class HandlerTest extends TestCase
 {
+    use BlackBox;
+
     public function testNothingHappensByDefaultWhenReceivingSignal()
     {
-        $handler = new Handler;
+        $handler = Handler::main();
 
         $this->fork();
 
@@ -22,18 +29,32 @@ class HandlerTest extends TestCase
 
     public function testAllListenersAreCalledInOrderOnSignal()
     {
-        $handlers = new Handler;
+        $handlers = Handler::main();
         $order = [];
         $count = 0;
 
         $this->fork();
 
         $this->assertNull($handlers->listen(Signal::child, function($signal) use (&$order, &$count): void {
+            static $handled = false;
+
+            if ($handled) {
+                return;
+            }
+
+            $handled = true;
             $this->assertSame(Signal::child, $signal);
             $order[] = 'first';
             ++$count;
         }));
         $handlers->listen(Signal::child, function($signal) use (&$order, &$count): void {
+            static $handled = false;
+
+            if ($handled) {
+                return;
+            }
+
+            $handled = true;
             $this->assertSame(Signal::child, $signal);
             $order[] = 'second';
             ++$count;
@@ -47,7 +68,7 @@ class HandlerTest extends TestCase
 
     public function testRemoveSignal()
     {
-        $handlers = new Handler;
+        $handlers = Handler::main();
         $order = [];
         $count = 0;
 
@@ -72,41 +93,10 @@ class HandlerTest extends TestCase
         $this->assertSame(['second'], $order);
     }
 
-    public function testListenersAddedAfterAResetAreNotCalled()
-    {
-        $wasAsync = \pcntl_async_signals();
-        $handlers = new Handler;
-        $order = [];
-        $count = 0;
-
-        $this->fork();
-
-        $listener = static function($signal) use (&$order, &$count): void {
-            $order[] = 'first';
-            ++$count;
-        };
-        $handlers->listen(Signal::child, $listener);
-
-        $this->assertNull($handlers->reset());
-        $this->assertSame($wasAsync, \pcntl_async_signals());
-
-        try {
-            $handlers->listen(Signal::child, $listener);
-            $this->fail('it should throw');
-        } catch (\Exception $e) {
-            $this->assertInstanceOf(\LogicException::class, $e);
-        }
-
-        \sleep(2); // wait for child to stop
-
-        $this->assertSame(0, $count);
-        $this->assertSame([], $order);
-    }
-
     public function testDefaultHandlerRestoredWhenAllListenersRemovedForASignal()
     {
         $wasAsync = \pcntl_async_signals();
-        $handlers = new Handler;
+        $handlers = Handler::main();
         $order = [];
         $count = 0;
 
@@ -125,6 +115,52 @@ class HandlerTest extends TestCase
 
         $this->assertSame(0, $count);
         $this->assertSame([], $order);
+    }
+
+    public function testAsyncHandlers(): BlackBox\Proof
+    {
+        return $this
+            ->forAll(Set::of(...Signal::cases()))
+            ->prove(function($signal) {
+                $main = Handler::main();
+                $interceptor = Interceptor::new();
+                $async = $main->async($interceptor);
+
+                $called = false;
+                $async->listen($signal, function($in) use ($signal, &$called) {
+                    $this->assertSame($signal, $in);
+                    $called = true;
+                });
+                $interceptor->dispatch($signal);
+
+                $this->assertTrue($called);
+            });
+    }
+
+    public function testRemovedAsyncListenersAreNotCalled(): BlackBox\Proof
+    {
+        return $this
+            ->forAll(Set::of(...Signal::cases()))
+            ->prove(function($signal) {
+                $main = Handler::main();
+                $interceptor = Interceptor::new();
+                $async = $main->async($interceptor);
+
+                $called = 0;
+                $listener = function($in) use ($signal, &$called) {
+                    $this->assertSame($signal, $in);
+                    ++$called;
+                };
+                $async->listen($signal, $listener);
+                $interceptor->dispatch($signal);
+
+                $this->assertSame(1, $called);
+
+                $async->remove($listener);
+                $interceptor->dispatch($signal);
+
+                $this->assertSame(1, $called);
+            });
     }
 
     private function fork(): void